Add alias specification to drone.ini
If an alias is specified for a port, it will be used as the port name
and the actual port name will be used as the description. Note that
this is only in the information sent to the controller. Drone still
continues to use the original name because pcap_XXX and other code
expect the actual port name