Implement Tx and increment port stats accordingly
No actual packets are crafted or transmitted. We just increment port
stats as per the configured stream. Only a single enabled stream Tx is
implemented.

NOTE: It is **highly likely** that the implemented tx/portStats logic is
wrong and doesn't work for all cases.

Per stream stats will not be supported in web demo.